+++ Libplanet.Explorer/Queries/TransactionQuery.cs
@@ -1,9 +1,14 @@
 #nullable disable
 using System;
+using Bencodex;
+using Bencodex.Types;
 using GraphQL;
 using GraphQL.Types;
 using Libplanet.Action;
+using Libplanet.Blockchain;
+using Libplanet.Crypto;
 using Libplanet.Explorer.GraphTypes;
+using Libplanet.Explorer.Interfaces;
 using Libplanet.Tx;
 
 namespace Libplanet.Explorer.Queries
@@ -11,8 +16,13 @@ namespace Libplanet.Explorer.Queries
     public class TransactionQuery<T> : ObjectGraphType
         where T : IAction, new()
     {
-        public TransactionQuery()
+        private static readonly Codec _codec = new Codec();
+        private readonly IBlockChainContext<T> _context;
+
+        public TransactionQuery(IBlockChainContext<T> context)
         {
+            _context = context ?? throw new ArgumentNullException(nameof(context));
+
             Field<NonNullGraphType<ListGraphType<NonNullGraphType<TransactionType<T>>>>>(
                 "transactions",
                 arguments: new QueryArguments(
@@ -104,6 +114,97 @@ namespace Libplanet.Explorer.Queries
                         ByteUtil.ParseHex(context.GetArgument<string>("id"))
                     );
                     return ExplorerQuery<T>.GetTransaction(id);
+                }
+            );
+
+            Field<NonNullGraphType<ByteStringType>>(
+                name: "unsignedTransaction",
+                arguments: new QueryArguments(
+                    new QueryArgument<NonNullGraphType<StringGraphType>>
+                    {
+                        Name = "publicKey",
+                        Description = "The hexadecimal string of public key for Transaction.",
+                    },
+                    new QueryArgument<NonNullGraphType<StringGraphType>>
+                    {
+                        Name = "plainValue",
+                        Description = "The hexadecimal string of plain value for Action.",
+                    },
+                    new QueryArgument<LongGraphType>
+                    {
+                        Name = "nonce",
+                        Description = "The nonce for Transaction.",
+                    }
+                ),
+                resolve: context =>
+                {
+                    BlockChain<T> chain = _context.BlockChain;
+                    string plainValueString = context.GetArgument<string>("plainValue");
+                    IValue plainValue = _codec.Decode(ByteUtil.ParseHex(plainValueString));
+                    var action = new T();
+                    action.LoadPlainValue(plainValue);
+
+                    var publicKey = new PublicKey(
+                        ByteUtil.ParseHex(context.GetArgument<string>("publicKey"))
+                    );
+                    Address signer = publicKey.ToAddress();
+                    long nonce = context.GetArgument<long?>("nonce") ??
+                        chain.GetNextTxNonce(signer);
+                    Transaction<T> unsignedTransaction =
+                        Transaction<T>.CreateUnsigned(
+                            nonce,
+                            publicKey,
+                            chain.Genesis.Hash,
+                            new[] { action }
+                        );
+                    return unsignedTransaction.Serialize(false);
+                }
+            );
+
+            Field<NonNullGraphType<StringGraphType>>(
+                name: "bindSignature",
+                #pragma warning disable MEN002
+                description: "Attach the given signature to the given transaction and return tx as hexadecimal",
+                #pragma warning restore MEN002
+                arguments: new QueryArguments(
+                    new QueryArgument<NonNullGraphType<StringGraphType>>
+                    {
+                        Name = "unsignedTransaction",
+                        #pragma warning disable MEN002
+                        Description = "The hexadecimal string of unsigned transaction to attach the given signature.",
+                        #pragma warning restore MEN002
+                    },
+                    new QueryArgument<NonNullGraphType<StringGraphType>>
+                    {
+                        Name = "signature",
+                        #pragma warning disable MEN002
+                        Description = "The hexadecimal string of the given unsigned transaction.",
+                        #pragma warning restore MEN002
+                    }
+                ),
+                resolve: context =>
+                {
+                    byte[] signature = ByteUtil.ParseHex(
+                        context.GetArgument<string>("signature")
+                    );
+                    Transaction<T> unsignedTransaction =
+                        Transaction<T>.Deserialize(
+                            ByteUtil.ParseHex(
+                                context.GetArgument<string>("unsignedTransaction")
+                            ),
+                            false
+                        );
+                    var signedTransaction = new Transaction<T>(
+                        unsignedTransaction.Nonce,
+                        unsignedTransaction.Signer,
+                        unsignedTransaction.PublicKey,
+                        unsignedTransaction.GenesisHash,
+                        unsignedTransaction.UpdatedAddresses,
+                        unsignedTransaction.Timestamp,
+                        unsignedTransaction.Actions,
+                        signature);
+
+                    return ByteUtil.Hex(signedTransaction.Serialize(true));
                 }
             );